+++ packages/gamut-templates/src/Notification/contstants.tsx
@@ -1,0 +1,14 @@
+import {
+  AlertIcon,
+  InformationalIcon,
+  CheckmarkIcon,
+  CloseIcon,
+} from '@codecademy/gamut';
+
+export const BANNER_CONFIG = {
+  alert: AlertIcon,
+  error: CloseIcon,
+  info: InformationalIcon,
+  announcement: AlertIcon,
+  success: CheckmarkIcon,
+};
