+++ packages/gamut-templates/src/Notification/types.ts
@@ -1,0 +1,12 @@
+export type BannerCTA = {
+  href?: string;
+  text: string;
+  onClick: () => void;
+};
+
+export type BannerTypes =
+  | 'alert'
+  | 'announcement'
+  | 'error'
+  | 'info'
+  | 'success';
